Split CategoryService to interface and implementation class
