+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/CategoryService.java
@@ -2,35 +2,13 @@ package com.avdo.spring.app.service;
 
 import com.avdo.spring.app.controller.dto.CreateCategoryDto;
 import com.avdo.spring.app.entity.CategoryEntity;
-import com.avdo.spring.app.repository.CategoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import com.avdo.spring.app.service.domain.request.CreateCategoryRequest;
 
 import java.util.List;
 
+public interface CategoryService {
 
-@Service
-public class CategoryService {
-
-    private final CategoryRepository categoryRepository;
-
-    @Autowired
-    public CategoryService(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
-    }
-
-    public CategoryEntity createCategory(CreateCategoryDto createCategoryDto) {
-        CategoryEntity category = new CategoryEntity();
-        category.setName(createCategoryDto.getName());
-        return categoryRepository.save(category);
-    }
-
-    public List<CategoryEntity> findAllCategories() {
-        return categoryRepository.findAll();
-    }
-
-    public CategoryEntity findCategoryByName(String name) {
-        return categoryRepository.findCategoryByName(name)
-                .orElseThrow();
-    }
+    List<CategoryEntity> findAllCategories();
+    CategoryEntity createCategory(CreateCategoryRequest createCategoryRequest);
+    CategoryEntity findCategoryByName(String name);
 }
